refactor: extract systemd code into _linux.go files with build tags
Split install/uninstall/status into shared logic + platform-specific
implementations behind //go:build linux tags. Update help text to
platform-neutral wording. Split systemd_test.go into per-command
_linux_test.go files.
